+++ src/ephemeral.rs
@@ -15,7 +15,9 @@ use crate::dalek::ephemeral as eph;
 use crate::sodium::ephemeral as eph;
 
 #[cfg(any(feature = "sodium", feature = "dalek"))]
-pub use eph::{derive_shared_secret, derive_shared_secret_pk, derive_shared_secret_sk};
+pub use eph::{
+    derive_shared_secret, derive_shared_secret_pk, derive_shared_secret_sk, sk_to_curve,
+};
 
 #[cfg(any(feature = "sodium", all(feature = "dalek", feature = "getrandom")))]
 pub use eph::generate_ephemeral_keypair;